refactor(value): move NumberDeserializer into number module, keep both serializer and deserializer pub(crate)
