+++ src/value/mod.rs
@@ -15,7 +15,8 @@ mod number;
 pub(crate) mod raw;
 
 pub use map::Map;
-pub use number::{Number, NumberSerializer, F32, F64};
+pub use number::{Number, F32, F64};
+pub(crate) use number::{NumberDeserializer, NumberSerializer};
 #[allow(clippy::useless_attribute, clippy::module_name_repetitions)]
 pub use raw::RawValue;
 
@@ -252,25 +253,6 @@ impl<'a, 'de> MapAccess<'de> for MapAccessor<'a> {
 
     fn size_hint(&self) -> Option<usize> {
         Some(self.items.len())
-    }
-}
-
-pub(crate) struct NumberDeserializer(pub(crate) Number);
-
-impl<'de> serde::de::Deserializer<'de> for NumberDeserializer {
-    type Error = crate::error::Error;
-
-    fn deserialize_any<V: serde::de::Visitor<'de>>(
-        self,
-        visitor: V,
-    ) -> crate::error::Result<V::Value> {
-        self.0.visit(visitor)
-    }
-
-    serde::forward_to_deserialize_any! {
-        bool i8 i16 i32 i64 i128 u8 u16 u32 u64 u128 f32 f64
-        char str string bytes byte_buf option unit unit_struct
-        newtype_struct seq tuple tuple_struct map struct enum identifier ignored_any
     }
 }
